pyproject.toml
[tool.poetry]
name = "code-rag"
version = "0.1.0"
description = ""
authors = ["sebastianbarrozo <sebastian.barrozo@etendo.software>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain-openai = "^0.1.7"
tiktoken = "^0.7.0"
langchain-chroma = "^0.1.1"
langchain = "^0.2.0"
gitpython = "^3.1.43"
langchain-community = "^0.2.0"
tree-sitter = "^0.20.4"
tree-sitter-languages = "^1.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
